pyproject.toml
[tool.poetry]
name = "hydra-optuna-sweeper-reborn"
version = "0.1.0"
description = "An updated Optuna Sweeper plugin for Hydra with pruning, dashboard, and callbacks support"
authors = []
license = "MIT"
readme = "README.md"
keywords = ["hydra", "optuna", "hyperparameter-optimization", "sweeper", "pruning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [
    { include = "hydra_plugins" }
]

[tool.poetry.dependencies]
python = ">=3.10"
hydra-core = ">=1.3.2"
optuna = ">=4.0.0"
omegaconf = ">=2.3.0"
optuna-dashboard = { version = ">=0.20.0", optional = true }

[tool.poetry.extras]
dashboard = ["optuna-dashboard"]

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.0"
pytest-cov = ">=7.0.0"
ruff = ">=0.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
